+++ crates/router/src/types/fraud_check.rs
@@ -1,11 +1,12 @@
-use common_utils::pii::Email;
+pub use hyperswitch_domain_models::router_request_types::fraud_check::{
+    FraudCheckCheckoutData, FraudCheckFulfillmentData, FraudCheckRecordReturnData,
+    FraudCheckSaleData, FraudCheckTransactionData, RefundMethod,
+};
 
 use crate::{
-    connector::signifyd::transformers::RefundMethod,
-    core::fraud_check::types::FrmFulfillmentRequest,
     pii::Serialize,
     services,
-    types::{self, api, storage_enums, ErrorResponse, ResponseId, RouterData},
+    types::{api, storage_enums, ErrorResponse, ResponseId, RouterData},
 };
 
 pub type FrmSaleRouterData = RouterData<api::Sale, FraudCheckSaleData, FraudCheckResponseData>;
@@ -13,13 +14,6 @@ pub type FrmSaleRouterData = RouterData<api::Sale, FraudCheckSaleData, FraudChec
 pub type FrmSaleType =
     dyn services::ConnectorIntegration<api::Sale, FraudCheckSaleData, FraudCheckResponseData>;
 
-#[derive(Debug, Clone)]
-pub struct FraudCheckSaleData {
-    pub amount: i64,
-    pub order_details: Option<Vec<api_models::payments::OrderDetailsWithAmount>>,
-    pub currency: Option<common_enums::Currency>,
-    pub email: Option<Email>,
-}
 #[derive(Debug, Clone)]
 pub struct FrmRouterData {
     pub merchant_id: String,
@@ -76,17 +70,6 @@ pub type FrmCheckoutType = dyn services::ConnectorIntegration<
     FraudCheckResponseData,
 >;
 
-#[derive(Debug, Clone)]
-pub struct FraudCheckCheckoutData {
-    pub amount: i64,
-    pub order_details: Option<Vec<api_models::payments::OrderDetailsWithAmount>>,
-    pub currency: Option<common_enums::Currency>,
-    pub browser_info: Option<types::BrowserInformation>,
-    pub payment_method_data: Option<api_models::payments::AdditionalPaymentData>,
-    pub email: Option<Email>,
-    pub gateway: Option<String>,
-}
-
 pub type FrmTransactionRouterData =
     RouterData<api::Transaction, FraudCheckTransactionData, FraudCheckResponseData>;
 
@@ -95,19 +78,6 @@ pub type FrmTransactionType = dyn services::ConnectorIntegration<
     FraudCheckTransactionData,
     FraudCheckResponseData,
 >;
-
-#[derive(Debug, Clone)]
-pub struct FraudCheckTransactionData {
-    pub amount: i64,
-    pub order_details: Option<Vec<api_models::payments::OrderDetailsWithAmount>>,
-    pub currency: Option<storage_enums::Currency>,
-    pub payment_method: Option<storage_enums::PaymentMethod>,
-    pub error_code: Option<String>,
-    pub error_message: Option<String>,
-    pub connector_transaction_id: Option<String>,
-    //The name of the payment gateway or financial institution that processed the transaction.
-    pub connector: Option<String>,
-}
 
 pub type FrmFulfillmentRouterData =
     RouterData<api::Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>;
@@ -125,18 +95,3 @@ pub type FrmRecordReturnType = dyn services::ConnectorIntegration<
     FraudCheckRecordReturnData,
     FraudCheckResponseData,
 >;
-
-#[derive(Debug, Clone)]
-pub struct FraudCheckFulfillmentData {
-    pub amount: i64,
-    pub order_details: Option<Vec<masking::Secret<serde_json::Value>>>,
-    pub fulfillment_req: FrmFulfillmentRequest,
-}
-
-#[derive(Debug, Clone)]
-pub struct FraudCheckRecordReturnData {
-    pub amount: i64,
-    pub currency: Option<storage_enums::Currency>,
-    pub refund_method: RefundMethod,
-    pub refund_transaction_id: Option<String>,
-}
